+++ packages/taro-rn-runner/src/index.ts
@@ -1,9 +1,9 @@
 import * as webpack from 'webpack'
-import { getSassLoaderOption } from '@tarojs/runner-utils'
 
-import { IBuildConfig, IOption } from './utils/types'
+import { IBuildConfig } from './utils/types'
 import { PARSE_AST_TYPE } from './utils/constants'
 import { printBuildError, bindProdLogger, bindDevLogger } from './utils/logHelper'
+import baseConf from './webpack/base.conf'
 import buildConf from './webpack/build.conf'
 
 const customizeChain = async (chain, modifyWebpackChainFunc: Function, customizeFunc: Function) => {
@@ -15,20 +15,13 @@ const customizeChain = async (chain, modifyWebpackChainFunc: Function, customize
   }
 }
 
-const makeConfig = async (buildConfig: IBuildConfig) => {
-  const sassLoaderOption: IOption = await getSassLoaderOption(buildConfig)
-  return {
-    ...buildConfig,
-    sassLoaderOption
-  }
-}
-
-export default async function build (appPath: string, config: IBuildConfig, mainBuilder) {
+export default async function build (appPath: string, config: IBuildConfig) {
   const mode = config.mode
-  const newConfig = await makeConfig(config)
+  const baseWebpackChain = baseConf(appPath)
   // config.webpackChain 自定义 Webpack 配置，接受函数形式的配置。
-  const webpackChain = buildConf(appPath, mode, config)
-  await customizeChain(webpackChain, newConfig.modifyWebpackChain, newConfig.webpackChain)
+  await customizeChain(baseWebpackChain, config.modifyWebpackChain, config.webpackChain)
+  const buildWebpackConf = buildConf(appPath, mode, config, baseWebpackChain)
+  const webpackChain = baseWebpackChain.merge(buildWebpackConf)
   const webpackConfig = webpackChain.toConfig()
   webpackConfig.stats = 'verbose'
   webpackConfig.profile = true